Did all the things (mostly)
Achieved pretty much all of the feature requests.

I speant a lot more time that I was supposed to. Really modeling the
domain in several tables added a lot of complexity, and I really didn't
know what I was doing with the react code that was provided as a
starting point.

I took the opportunity to brush up on functional coponents in React, and
had a pretty good time chasing all the bits around :)

Some of the minitest tests are skipped because there is some bug causing
extra requests to be made, haven't been able to track down where that's
coming from.

Definitely a good bit more test coverage is warranted, but I tried to
add coverage the the most interesting bits (or bits that were causing
bugs)

I removed the sleeps from the two external service classes, That could
should really be pushed to a job queue, such as Sidekiq, but that was a
bridge too far for the exercise :)

I appoligize for my JS lol. I really should have linted everything, but
again, was a bit over time.